pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xauusd-trading"
version = "0.1.0"
description = "NN-based XAUUSD futures trading system"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1.0",
    "onnxruntime>=1.16.0",
    "onnx>=1.15.0",
    "polars>=0.20.0",
    "numpy>=1.24.0",
    "duckdb>=0.9.0",
    "transformers>=4.35.0",
    "stable-baselines3>=2.2.0",
    "d3rlpy>=2.3.0",
    "gymnasium>=0.29.0",
    "pandas-ta>=0.3.14b",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "prometheus-client>=0.19.0",
    "python-telegram-bot>=20.7",
    "requests>=2.31.0",
    "pyarrow>=14.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.8.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
mt5 = ["MetaTrader5>=5.0.45"]
ibkr = ["ib_insync>=0.9.86"]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
